Add files via upload
The project title is “CRICKET ALERT’S”. By this project we can know the live cricket score of the match by enter the details( eg: team names, match type…) of the match. The live score will be send through the whatsapp messenger app. 

To get the live score we have used the “CRICAPI” website. In this website we will get the details of the cricket matches live and also upcoming. There will be different API’s in this website.
List of API’s:
                    └ Current Matches
                    └ eCricScore
                    └ Series Search
                    └ Series List
                    └ Matches List
                    └ Players List
                    └ Players Search
And there are many API’s in this website. But from the above list of API’s we have used only two i.e, └ Current Matches and └ eCricScore.

Up to now we had seen about how live cricket score is generated. And now we will see how we will get the score to the whatsapp messenger app.
For this we had used the “TWILIO” it is an American company based in San Francisco, California, which provides programmable communication tools for making and receiving phone calls, sending and receiving text messages, and performing other communication functions using its web service APIs.